requirement.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-rest-framework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# File Handling
Pillow>=10.0.0
python-magic>=0.4.27

# Machine Learning (for inference)
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Computer Vision
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# Security
bcrypt>=4.1.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0